requirements.txt
jupyterlab==4.0.2
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
opencv-python==4.8.0.74
ultralytics-8.0.131
# torch-model-archiver==0.8.1
torch-model-archiver==0.6.0
torch-workflow-archiver==0.2.9
fastapi==0.100.0
uvicorn==0.23.0
python-multipart==0.0.6
omegaconf==2.3.0
dependency_injector==4.41.0
